requirements.txt
torch>=2.0.0
torchvision>=0.15.0
sentence-transformers>=2.2.0
transformers>=4.35.0
open-clip-torch>=2.23.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
chromadb>=0.4.0
ollama>=0.1.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=7.4.0